+++ src/main/java/org/greencodeinitiative/creedengo/python/PythonRuleRepository.java
@@ -47,6 +47,7 @@ public class PythonRuleRepository implements RulesDefinition, PythonCustomRuleRe
             DictionaryItemsUnused.class,
             AvoidCSVFormat.class,
             AvoidIterativeMatrixOperations.class,
+            AvoidNonPinnedMemoryForDataloaders.class,
             AvoidConvBiasBeforeBatchNorm.class
     );